Code corrections
The stack was always allocating 10 elements, regardless of stack_size_allocated, changed that. Also, the if condition to reallocate bits of the stack (whenever it needed to be increased) had a legacy code portion based on byte size allocation in stack_size_allocated, which has been changed to number of elements. Also, it was allocating one more than needed (>= -> >). Also tested the code for different types of input elements, and it seems to be working.

+++ simplemain.c
@@ -8,10 +8,10 @@ int main()
 {
                
         
-        create_stack(sizeof(uint8_t));
+        create_stack(sizeof(uint16_t));
         printf("%u\n",check_stack_is_empty());
 
-        uint8_t data_test = 3;
+        uint16_t data_test = 3;
         stack_push((void*) &data_test);
 
         data_test = 20;
@@ -21,13 +21,13 @@ int main()
         stack_push((void*) &data_test);
         
 
-        printf("%u\n", *((uint8_t*)check_stack_top()));
+        printf("%u\n", *((uint16_t*)check_stack_top()));
         stack_pop();
 
-        printf("%u\n", *((uint8_t*)check_stack_top()));
+        printf("%u\n", *((uint16_t*)check_stack_top()));
         stack_pop();
 
-        printf("%u\n", *((uint8_t*)check_stack_top()));
+        printf("%u\n", *((uint16_t*)check_stack_top()));
 
 
         printf("%u\n",check_stack_size());
